requirements.txt
beautifulsoup4
markdownify
playwright
requests
rich
openai
python-dotenv
google-generativeai
Pillow
tqdm
scikit-learn
# fastapi
# uvicorn
# python-dotenv
# aiohttp
numpy
# pydantic
beautifulsoup4
# html2text
# tqdm
# markdown
# python-multipart
# setuptools